use type size_t instead of int since the 3rd argument of sc_format_asn1_entry() is void * using int will fail on a 64-bits platform
Closes Ticket #176


git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@3509 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/tools/pkcs15-tool.c
@@ -1421,7 +1421,7 @@ static int pem_encode(int alg_id, sc_pkcs15_der_t *key, sc_pkcs15_der_t *out)
 	struct sc_asn1_entry	asn1_pem_key[2],
 				asn1_pem_key_items[3];
 	struct sc_algorithm_id algorithm;
-	int key_len;
+	size_t key_len;
 
 	memset(&algorithm, 0, sizeof(algorithm));
 	algorithm.algorithm = alg_id;
